+++ src/ace_button/ButtonConfig.h
@@ -26,6 +26,17 @@ SOFTWARE.
 #define ACE_BUTTON_BUTTON_CONFIG_H
 
 #include <Arduino.h>
+#include "IEventHandler.h"
+
+// https://stackoverflow.com/questions/295120
+#if defined(__GNUC__) || defined(__clang__)
+  #define DEPRECATED __attribute__((deprecated))
+#elif defined(_MSC_VER)
+  #define DEPRECATED __declspec(deprecated)
+#else
+  #pragma message("WARNING: You need to implement DEPRECATED for this compiler")
+  #define DEPRECATED
+#endif
 
 namespace ace_button {
 
@@ -134,6 +145,12 @@ class ButtonConfig {
      * kEventDoubleClicked.
      */
     static const FeatureFlagType kFeatureSuppressClickBeforeDoubleClick = 0x100;
+
+    /**
+     * Internal flag to indicate that mEventHandler is an IEventHandler object
+     * pointer instead of an EventHandler function pointer.
+     */
+    static const FeatureFlagType kInternalFeatureIEventHandler = 0x8000;
 
     /**
      * Convenience flag to suppress all suppressions. Calling
@@ -289,26 +306,68 @@ class ButtonConfig {
     }
 
     /**
-     * Disable all features. Useful when the ButtonConfig is reused in different
-     * configurations. Also useful for testing.
+     * Disable all (externally visible) features. Useful when the ButtonConfig
+     * is reused in different configurations. Also useful for testing. Internal
+     * feature flags (e.g. kInternalFeatureIEventHandler) are *not* cleared.
      */
     void resetFeatures() {
-      mFeatureFlags = 0;
+      // NOTE: If any additional kInternalFeatureXxx flag is added, it must be
+      // added here like this:
+      // mFeatureFlags &= (kInternalFeatureIEventHandler | kInternalFeatureXxx)
+      mFeatureFlags &= kInternalFeatureIEventHandler;
     }
 
     // EventHandler
 
-    /** Return the eventHandler. */
-    EventHandler getEventHandler() const {
-      return mEventHandler;
+    /**
+     * Return the eventHandler function pointer. This is meant to be an
+     * internal method.
+     *
+     * Deprecated as of v1.6 because the event handler can now be either a
+     * function pointer or an object pointer. AceButton class now calls
+     * dispatchEvent() which correctly handles both cases. Application code
+     * should never need to retrieve the event handler directly.
+     */
+    EventHandler getEventHandler() const DEPRECATED {
+      return reinterpret_cast<EventHandler>(mEventHandler);
     }
 
     /**
-     * Install the event handler. The event handler must be defined for the
-     * AceButton to be useful.
+     * Dispatch the event to the handler. This is meant to be an internal
+     * method.
+     */
+    void dispatchEvent(AceButton* button, uint8_t eventType,
+        uint8_t buttonState) const {
+
+      if (! mEventHandler) return;
+
+      if (isFeature(kInternalFeatureIEventHandler)) {
+        IEventHandler* eventHandler =
+            reinterpret_cast<IEventHandler*>(mEventHandler);
+        eventHandler->handleEvent(button, eventType, buttonState);
+      } else {
+        EventHandler eventHandler =
+            reinterpret_cast<EventHandler>(mEventHandler);
+        eventHandler(button, eventType, buttonState);
+      }
+    }
+
+    /**
+     * Install the EventHandler function pointer. The event handler must be
+     * defined for the AceButton to be useful.
      */
     void setEventHandler(EventHandler eventHandler) {
+      mEventHandler = reinterpret_cast<void*>(eventHandler);
+      clearFeature(kInternalFeatureIEventHandler);
+    }
+
+    /**
+     * Install the IEventHandler object pointer. The event handler must be
+     * defined for the AceButton to be useful.
+     */
+    void setIEventHandler(IEventHandler* eventHandler) {
       mEventHandler = eventHandler;
+      setFeature(kInternalFeatureIEventHandler);
     }
 
     /**
@@ -331,7 +390,7 @@ class ButtonConfig {
     ButtonConfig& operator=(const ButtonConfig&) = delete;
 
     /** The event handler for all buttons associated with this ButtonConfig. */
-    EventHandler mEventHandler = nullptr;
+    void* mEventHandler = nullptr;
 
     /** A bit mask flag that activates certain features. */
     FeatureFlagType mFeatureFlags = 0;